# Convert SdbData types to readonly structs