--- a/src/DocumentFormat.OpenXml/Validation/Schema/SdbClassIdToSchemaTypeIndex.cs
+++ b/src/DocumentFormat.OpenXml/Validation/Schema/SdbClassIdToSchemaTypeIndex.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-/**********************************************************
- * Define data struct for schema constraint binary database
- **********************************************************/
-
 using System.Diagnostics;
 
-using SdbIndex = System.UInt16;
+using static DocumentFormat.OpenXml.Validation.Schema.SdbData;
 
 namespace DocumentFormat.OpenXml.Validation.Schema
 {
@@ -15,89 +11,70 @@ namespace DocumentFormat.OpenXml.Validation.Schema
     /// Map OpenXmlElement class ID to schema type in schema.
     /// </summary>
     [DebuggerDisplay("ClassId={ClassId}")]
-    internal class SdbClassIdToSchemaTypeIndex : SdbData
+    internal readonly struct SdbClassIdToSchemaTypeIndex
     {
-        /// <summary>
-        /// Start ID of class ID.
-        /// </summary>
-        public const ushort StartClassId = 10001;
-
-        public const ushort InvalidSchemaTypeIndex = ushort.MaxValue;
+        public static readonly int TypeSize = sizeof(ushort) * 2;
 
         /// <summary>
-        /// Gets or sets class ID (Element Type ID).
+        /// Initializes an instance of <see cref="SdbClassIdToSchemaTypeIndex"/>
         /// </summary>
-        public ushort ClassId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the index of the schema type in the SdbSchemaType data array.
-        /// </summary>
-        public ushort SchemaTypeIndex { get; set; }
-
-        /// <summary>
-        /// Initializes a new instance of the SdbClassIdToSchemaTypeIndex.
-        /// </summary>
-        public SdbClassIdToSchemaTypeIndex()
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the SdbClassIdToSchemaTypeIndex.
-        /// </summary>
-        /// <param name="classId"></param>
-        /// <param name="schemaTypeIndex"></param>
-        public SdbClassIdToSchemaTypeIndex(ushort classId, ushort schemaTypeIndex)
+        public SdbClassIdToSchemaTypeIndex(ushort classId, ushort typeIndex)
         {
             ClassId = classId;
-            SchemaTypeIndex = schemaTypeIndex;
+            SchemaTypeIndex = typeIndex;
         }
+
+        /// <summary>
+        /// Initializes an instance of <see cref="SdbClassIdToSchemaTypeIndex"/> that deserializes from data
+        /// </summary>
+        /// <remarks>
+        /// The order of <see cref="SdbClassIdToSchemaTypeIndex(byte[], int)"/> and <see cref="Serialize"/> must remain
+        /// in sync to facilitate serialization and deserialization of binary data
+        /// </remarks>
+        private SdbClassIdToSchemaTypeIndex(byte[] data, int startIndex)
+        {
+            ClassId = LoadSdbIndex(data, ref startIndex);
+            SchemaTypeIndex = LoadSdbIndex(data, ref startIndex);
+        }
+
+        /// <summary>
+        /// Gets class ID (Element Type ID).
+        /// </summary>
+        public ushort ClassId { get; }
+
+        /// <summary>
+        /// Gets the index of the schema type in the SdbSchemaType data array.
+        /// </summary>
+        public ushort SchemaTypeIndex { get; }
+
+#if DEBUG
+        public const ushort StartClassId = 10001;
 
         /// <summary>
         /// Return the index of the data in the data array. The data array is sorted by the class ID and the class ID is continuous.
         /// </summary>
-        /// <param name="classId"></param>
-        /// <returns></returns>
         public static ushort ArrayIndexFromClassId(ushort classId)
         {
             Debug.Assert(classId >= StartClassId);
             return (ushort)(classId - StartClassId);
         }
+#endif
+
+        public static SdbClassIdToSchemaTypeIndex Deserialize(byte[] data, int startIndex) => new SdbClassIdToSchemaTypeIndex(data, startIndex);
 
         /// <summary>
-        /// Gets the size in bytes of this data structure.
+        /// Serializes the instance data to a byte array
         /// </summary>
-        public static int TypeSize => sizeof(ushort) * 2;
-
-        #region Override SdbData Members
-
-        /// <summary>
-        /// Gets the size in bytes of this data structure.
-        /// </summary>
-        public override int DataSize => TypeSize;
-
-        /// <summary>
-        /// Serialize the data into byte data.
-        /// </summary>
-        /// <returns>Byte data.</returns>
-        public override byte[] GetBytes()
+        /// <remarks>
+        /// The order of <see cref="SdbClassIdToSchemaTypeIndex(byte[], int)"/> and <see cref="Serialize"/> must remain
+        /// in sync to facilitate serialization and deserialization of binary data
+        /// </remarks>
+        public byte[] Serialize()
         {
-            // !!!!Caution: keep the order of the following code lines!!!!
-            return GetBytes(ClassId.Bytes(),
-                                 SchemaTypeIndex.Bytes());
+            return GetBytes(
+                TypeSize,
+                ClassId.Bytes(),
+                SchemaTypeIndex.Bytes());
         }
-
-        /// <summary>
-        /// Deserialize the data from byte data.
-        /// </summary>
-        /// <param name="value">The byte data.</param>
-        /// <param name="startIndex">The offset the data begins at.</param>
-        public override void LoadFromBytes(byte[] value, int startIndex)
-        {
-            // !!!!Caution: keep the order of the following code lines!!!!
-            ClassId = LoadSdbIndex(value, ref startIndex);
-            SchemaTypeIndex = LoadSdbIndex(value, ref startIndex);
-        }
-
-        #endregion
     }
 }