… src/DocumentFormat.OpenXml/Validation/Schema/BinarySdbSchemaDatas.cs
… src/DocumentFormat.OpenXml/Validation/Schema/SdbParticleChildrenIndex.cs
… src/DocumentFormat.OpenXml/Validation/Schema/SdbParticleConstraint.cs
… src/DocumentFormat.OpenXml/Validation/Schema/SdbSchemaData.Validation.cs
… src/DocumentFormat.OpenXml/Validation/Schema/SdbSchemaData.cs
… src/DocumentFormat.OpenXml/Validation/Schema/SdbSchemaType.cs
… src/DocumentFormat.OpenXml/Validation/Schema/ValueExtensionMethods.cs
… test/DocumentFormat.OpenXml.Tests/Schema/SdbDataTests.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/AllParticleValidatorTest.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/AnyParticleValidatorTest.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/ChoiceParticleValidatorTest.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/CompositeParticleValidatorTest.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/GroupParticleValidatorTest.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/SdbSchemaDatasTest.cs
… test/DocumentFormat.OpenXml.Tests/ofapiTest/SequenceParticleValidatorTest.cs