This is for performance (also cleans up some old code) that reduces the allocations used when loading the schema constraint data. Previously new objects were loaded for each of the items (which are thousands). This changes it to use structs instead of objects, which greatly reduces the memory allocated.

master:
```
       Method |         Mean |         Error |        StdDev |     Gen 0 |     Gen 1 |    Gen 2 |   Allocated |
------------- |-------------:|--------------:|--------------:|----------:|----------:|---------:|------------:|
       Create |    160.83 us |     3.1316 us |     4.1806 us |   44.1895 |         - |        - |    68.01 KB |
 CreateNoSave |     49.63 us |     0.6315 us |     0.5907 us |   16.5405 |         - |        - |    25.42 KB |
     ReadFile |  2,040.06 us |    31.1776 us |    29.1635 us |  312.5000 |   78.1250 |        - |   730.92 KB |
   Validation | 86,703.67 us | 1,587.5713 us | 1,485.0151 us | 4750.0000 | 1000.0000 | 250.0000 | 16155.36 KB |
     LoadData | 18,744.22 us |   367.6242 us |   393.3538 us |  531.2500 |  218.7500 |  62.5000 |  2988.81 KB |
```

with change:

```
       Method |         Mean |       Error |      StdDev |     Gen 0 |     Gen 1 |    Gen 2 |   Allocated |
------------- |-------------:|------------:|------------:|----------:|----------:|---------:|------------:|
       Create |    156.42 us |   1.7637 us |   1.6498 us |   44.1895 |         - |        - |    68.01 KB |
 CreateNoSave |     50.22 us |   0.9710 us |   0.9083 us |   16.5405 |         - |        - |    25.42 KB |
     ReadFile |  2,013.20 us |  19.1713 us |  17.9329 us |  312.5000 |   78.1250 |        - |   730.92 KB |
   Validation | 86,429.49 us | 786.2571 us | 656.5602 us | 4562.5000 | 1062.5000 | 312.5000 | 16015.66 KB |
     LoadData | 10,145.91 us | 171.3562 us | 160.2867 us |  328.1250 |  171.8750 |  46.8750 |  1874.71 KB |
```

There is also a perf gain in the time it takes to load all the data by about 43%, and a reduction of allocations by 38%